Add maintainer sync between rdo and rpmfactory
It is possible to pre-provision core/ptl users on a project created
on rpmfactory according to the maintainers list declared in rdo.
It is assumed that users will authenticate later on on rpmfactory through
Github's OAuth, and that if the user cannot be found by her email on
Github, then the 3rd party auth will acknowledge the maintainer email.

Change-Id: I80f5e1843996882c1626719b70cd42ac3f138e82