Add rpm-liberty/rpm-kilo/rdo-kilo branch to distgit sync
This change updates the sync_repo command so that
it also fetchs rpm-liberty/rpm-kilo/rdo-kilo branch.
If the branch doesn't exists upstream it simply skip,
and if the branch isn't created locally it does the
initial import.

This patch also try to support (for client, lib, and
None type) where distgit targets pkgs.fedoraproject.org:master
for rdo-liberty to check if mdistgit (github) own a rdo-liberty
branch. If this is the case rdo-liberty branch will be re-sync
from mdistgit(github):rdo-liberty.

Related-To: #1238

Change-Id: I1955b15eb6361c3bddd944e9b44f4c0747f7b80c