Add command to setup Github replication
This patch adds the command "replicate" to sfrdo to automatically setup
replication to Github. For example:

    sfrdo replicate --name swift --token <Github token>
    sfrdo replicate --name swift-distgit --token <Github token>

Repositories on Github are forked using their upstream or master-distgit
information from rdoinfo, and renamed if required.

Currently a rpeository name is required to setup replication. A follow up patch
will also allow to setup replication for all found repositories.

Change-Id: I1af763831e5b202e1f937e77e907c1d764163fb7
Depends-On: I775f0d310159da7b66c54ef0187effcb7cdcaef9